feat(code-example-insights): refactor evaluation results and diff statistics to use getters in CommunityCourseStageSolutionModel
